+++ backends-clickhouse/src/main/scala/org/apache/gluten/backendsapi/clickhouse/CHBackend.scala
@@ -405,4 +405,6 @@ object CHBackendSettings extends BackendSettingsApi with Logging {
         }
       }
   }
+
+  override def supportNativeInputFileRelatedExpr(): Boolean = true
 }
