--- cpp-ch/local-engine/Common/common.cpp
@@ -1,36 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-#include <Interpreters/Context.h>
-#include <Parser/SerializedPlanParser.h>
-
-#include <filesystem>
-
-using namespace DB;
-
-#ifdef __cplusplus
-extern "C" {
-#endif
-
-bool executorHasNext(char * executor_address)
-{
-    local_engine::LocalExecutor * executor = reinterpret_cast<local_engine::LocalExecutor *>(executor_address);
-    return executor->hasNext();
-}
-
-#ifdef __cplusplus
-}
-#endif
