+++ cpp-ch/local-engine/Parser/SerializedPlanParser.cpp
@@ -59,13 +59,13 @@
 #include <Parser/WriteRelParser.h>
 #include <Parsers/ASTIdentifier.h>
 #include <Parsers/ExpressionListParsers.h>
+#include <Parser/LocalExecutor.h>
 #include <Processors/QueryPlan/AggregatingStep.h>
 #include <Processors/QueryPlan/ExpressionStep.h>
 #include <Processors/QueryPlan/LimitStep.h>
 #include <Processors/QueryPlan/Optimizations/QueryPlanOptimizationSettings.h>
 #include <Processors/QueryPlan/QueryPlan.h>
 #include <Processors/QueryPlan/ReadFromPreparedSource.h>
-#include <Processors/Transforms/AggregatingTransform.h>
 #include <QueryPipeline/Pipe.h>
 #include <QueryPipeline/QueryPipelineBuilder.h>
 #include <QueryPipeline/printPipeline.h>
@@ -82,10 +82,9 @@
 #include <Common/GlutenConfig.h>
 #include <Common/JNIUtils.h>
 #include <Common/logger_useful.h>
-#include <Common/QueryContext.h>
 #include <Common/typeid_cast.h>
 #include <Processors/Executors/PipelineExecutor.h>
-#include <Processors/Executors/PullingAsyncPipelineExecutor.h>
+
 
 namespace DB
 {
@@ -300,11 +299,11 @@ QueryPlanStepPtr SerializedPlanParser::parseReadRealWithJavaIter(const substrait
 
     GET_JNIENV(env)
     SCOPE_EXIT({CLEAN_JNIENV});
-    auto * first_block = SourceFromJavaIter::peekBlock(env, input_iter);
+    auto first_block = SourceFromJavaIter::peekBlock(env, input_iter);
 
     /// Try to decide header from the first block read from Java iterator. Thus AggregateFunction with parameters has more precise types.
-    auto header = first_block ? first_block->cloneEmpty() : TypeParser::buildBlockFromNamedStruct(rel.base_schema());
-    auto source = std::make_shared<SourceFromJavaIter>(context, std::move(header), input_iter, materialize_input, first_block);
+    auto header = first_block.has_value() ? first_block->cloneEmpty() : TypeParser::buildBlockFromNamedStruct(rel.base_schema());
+    auto source = std::make_shared<SourceFromJavaIter>(context, std::move(header), input_iter, materialize_input, std::move(first_block));
 
     QueryPlanStepPtr source_step = std::make_unique<ReadFromPreparedSource>(Pipe(source));
     source_step->setStepDescription("Read From Java Iter");
@@ -432,6 +431,9 @@ QueryPlanPtr SerializedPlanParser::parse(const substrait::Plan & plan)
 
     return query_plan;
 }
+
+std::unique_ptr<LocalExecutor> SerializedPlanParser::createExecutor(const substrait::Plan & plan)
+{ return createExecutor(parse(plan), plan); }
 
 QueryPlanPtr SerializedPlanParser::parseOp(const substrait::Rel & rel, std::list<const substrait::Rel *> & rel_stack)
 {
@@ -1560,139 +1562,6 @@ void SerializedPlanParser::wrapNullable(
         actions_dag.addOrReplaceInOutputs(*node);
         nullable_measure_names[item] = node->result_name;
     }
-}
-
-LocalExecutor::~LocalExecutor()
-{
-    if (dump_pipeline)
-        LOG_INFO(&Poco::Logger::get("LocalExecutor"), "Dump pipeline:\n{}", dumpPipeline());
-
-    if (spark_buffer)
-    {
-        ch_column_to_spark_row->freeMem(spark_buffer->address, spark_buffer->size);
-        spark_buffer.reset();
-    }
-}
-
-std::unique_ptr<SparkRowInfo> LocalExecutor::writeBlockToSparkRow(const Block & block) const
-{
-    return ch_column_to_spark_row->convertCHColumnToSparkRow(block);
-}
-
-void LocalExecutor::initPullingPipelineExecutor()
-{
-    if (!executor)
-    {
-        query_pipeline = QueryPipelineBuilder::getPipeline(std::move(*query_pipeline_builder));
-        query_pipeline.setNumThreads(1);
-        executor = std::make_unique<PullingAsyncPipelineExecutor>(query_pipeline);
-    }
-}
-
-bool LocalExecutor::hasNext()
-{
-    initPullingPipelineExecutor();
-    size_t columns = currentBlock().columns();
-    if (columns == 0 || isConsumed())
-    {
-        auto empty_block = header.cloneEmpty();
-        setCurrentBlock(empty_block);
-        bool has_next = executor->pull(currentBlock());
-        produce();
-        return has_next;
-    }
-    return true;
-}
-
-SparkRowInfoPtr LocalExecutor::next()
-{
-    checkNextValid();
-    SparkRowInfoPtr row_info = writeBlockToSparkRow(currentBlock());
-    consume();
-    if (spark_buffer)
-    {
-        ch_column_to_spark_row->freeMem(spark_buffer->address, spark_buffer->size);
-        spark_buffer.reset();
-    }
-    spark_buffer = std::make_unique<SparkBuffer>();
-    spark_buffer->address = row_info->getBufferAddress();
-    spark_buffer->size = row_info->getTotalBytes();
-    return row_info;
-}
-
-Block * LocalExecutor::nextColumnar()
-{
-    checkNextValid();
-    Block * columnar_batch;
-    if (currentBlock().columns() > 0)
-    {
-        columnar_batch = &currentBlock();
-    }
-    else
-    {
-        auto empty_block = header.cloneEmpty();
-        setCurrentBlock(empty_block);
-        columnar_batch = &currentBlock();
-    }
-    consume();
-    return columnar_batch;
-}
-
-void LocalExecutor::cancel()
-{
-    if (executor)
-        executor->cancel();
-    if (push_executor)
-        push_executor->cancel();
-}
-
-void LocalExecutor::execute()
-{
-    chassert(query_pipeline_builder || external_pipeline_builder);
-    if (external_pipeline_builder)
-        push_executor = external_pipeline_builder->execute();
-    else
-        push_executor = query_pipeline_builder->execute();
-    push_executor->execute(local_engine::QueryContextManager::instance().currentQueryContext()->getSettingsRef().max_threads, false);
-}
-
-Block LocalExecutor::getHeader()
-{
-    return header;
-}
-
-LocalExecutor::LocalExecutor(QueryPlanPtr query_plan, QueryPipelineBuilderPtr pipeline_builder, bool dump_pipeline_)
-    : query_pipeline_builder(std::move(pipeline_builder))
-    , header(query_plan->getCurrentDataStream().header.cloneEmpty())
-    , dump_pipeline(dump_pipeline_)
-    , ch_column_to_spark_row(std::make_unique<CHColumnToSparkRow>())
-    , current_query_plan(std::move(query_plan))
-{
-    chassert(!current_executor);
-    current_executor = this;
-}
-thread_local LocalExecutor * LocalExecutor::current_executor = nullptr;
-std::string LocalExecutor::dumpPipeline() const
-{
-    const auto & processors = query_pipeline.getProcessors();
-    for (auto & processor : processors)
-    {
-        WriteBufferFromOwnString buffer;
-        auto data_stats = processor->getProcessorDataStats();
-        buffer << "(";
-        buffer << "\nexecution time: " << processor->getElapsedNs() / 1000U << " us.";
-        buffer << "\ninput wait time: " << processor->getInputWaitElapsedNs() / 1000U << " us.";
-        buffer << "\noutput wait time: " << processor->getOutputWaitElapsedNs() / 1000U << " us.";
-        buffer << "\ninput rows: " << data_stats.input_rows;
-        buffer << "\ninput bytes: " << data_stats.input_bytes;
-        buffer << "\noutput rows: " << data_stats.output_rows;
-        buffer << "\noutput bytes: " << data_stats.output_bytes;
-        buffer << ")";
-        processor->setDescription(buffer.str());
-    }
-    WriteBufferFromOwnString out;
-    printPipeline(processors, out);
-    return out.str();
 }
 
 NonNullableColumnsResolver::NonNullableColumnsResolver(
