+++ cpp-ch/local-engine/Parser/SerializedPlanParser.h
@@ -23,15 +23,12 @@
 #include <Interpreters/Aggregator.h>
 #include <Parser/CHColumnToSparkRow.h>
 #include <Parser/RelMetric.h>
-#include <Processors/Executors/PullingAsyncPipelineExecutor.h>
 #include <Processors/QueryPlan/ISourceStep.h>
 #include <Processors/QueryPlan/QueryPlan.h>
 #include <Storages/MergeTree/SparkStorageMergeTree.h>
 #include <Storages/SourceFromJavaIter.h>
 #include <base/types.h>
 #include <substrait/plan.pb.h>
-#include <Common/BlockIterator.h>
-#include <QueryPipeline/QueryPipelineBuilder.h>
 
 namespace local_engine
 {
@@ -88,7 +85,7 @@ public:
 
     /// visible for UT
     DB::QueryPlanPtr parse(const substrait::Plan & plan);
-    std::unique_ptr<LocalExecutor> createExecutor(const substrait::Plan & plan) { return createExecutor(parse(plan), plan); }
+    std::unique_ptr<LocalExecutor> createExecutor(const substrait::Plan & plan);
     DB::QueryPipelineBuilderPtr buildQueryPipeline(DB::QueryPlan & query_plan);
     ///
     std::unique_ptr<LocalExecutor> createExecutor(const std::string_view plan);
@@ -195,69 +192,6 @@ private:
 public:
     const ActionsDAG::Node * addColumn(DB::ActionsDAG & actions_dag, const DataTypePtr & type, const Field & field);
 };
-
-struct SparkBuffer
-{
-    char * address;
-    size_t size;
-};
-
-class LocalExecutor : public BlockIterator
-{
-public:
-    static LocalExecutor * getCurrentExecutor() { return current_executor; }
-    static void resetCurrentExecutor() { current_executor = nullptr; }
-    LocalExecutor(QueryPlanPtr query_plan, QueryPipelineBuilderPtr pipeline, bool dump_pipeline_ = false);
-    ~LocalExecutor();
-
-    SparkRowInfoPtr next();
-    Block * nextColumnar();
-    bool hasNext();
-    // When a fallback occurs, hasNext will be called to trigger the initialization of the pulling executor
-    bool initByPulling()
-    {
-        return executor.get();
-    }
-
-    /// Stop execution, used when task receives shutdown command or executor receives SIGTERM signal
-    void cancel();
-    void setSinks(std::function<void(QueryPipelineBuilder &)> setter)
-    {
-        setter(*query_pipeline_builder);
-    }
-    // set shuffle write pipeline for fallback
-    void setExternalPipelineBuilder(QueryPipelineBuilderPtr builder)
-    {
-        external_pipeline_builder = std::move(builder);
-    }
-    void execute();
-    Block getHeader();
-    RelMetricPtr getMetric() const { return metric; }
-    void setMetric(const RelMetricPtr & metric_) { metric = metric_; }
-    void setExtraPlanHolder(std::vector<QueryPlanPtr> & extra_plan_holder_) { extra_plan_holder = std::move(extra_plan_holder_); }
-
-private:
-    static thread_local LocalExecutor * current_executor;
-    std::unique_ptr<SparkRowInfo> writeBlockToSparkRow(const DB::Block & block) const;
-    void initPullingPipelineExecutor();
-    /// Dump processor runtime information to log
-    std::string dumpPipeline() const;
-
-    QueryPipelineBuilderPtr query_pipeline_builder;
-    // final shuffle write pipeline for fallback
-    QueryPipelineBuilderPtr external_pipeline_builder = nullptr;
-    QueryPipeline query_pipeline;
-    std::unique_ptr<DB::PullingAsyncPipelineExecutor> executor = nullptr;
-    PipelineExecutorPtr push_executor = nullptr;
-    Block header;
-    bool dump_pipeline;
-    std::unique_ptr<CHColumnToSparkRow> ch_column_to_spark_row;
-    std::unique_ptr<SparkBuffer> spark_buffer;
-    QueryPlanPtr current_query_plan;
-    RelMetricPtr metric;
-    std::vector<QueryPlanPtr> extra_plan_holder;
-};
-
 
 class ASTParser
 {
