+++ cpp-ch/local-engine/Parser/SparkRowToCHColumn.cpp
@@ -29,6 +29,7 @@
 #include <Functions/FunctionHelpers.h>
 #include <Common/CHUtil.h>
 #include <Common/Exception.h>
+#include <Common/DebugUtils.h>
 
 namespace DB
 {
@@ -124,6 +125,8 @@ Block * SparkRowToCHColumn::getBlock(SparkRowToCHColumnHelper & helper)
         ColumnWithTypeAndName named_col(col, uint8_ty, "__anonymous_col__");
         block->insert(named_col);
     }
+    std::cout << "spark row to ch column:\n";
+    debug::headBlock(*block);
     return block;
 }
